#3465: moved evertything to pkg
