+++ secret-service/pkg/backend/fake/secretbackend_mock.go
@@ -4,7 +4,7 @@
 package fake
 
 import (
-	"github.com/keptn/keptn/secret-service/internal/backend"
+	"github.com/keptn/keptn/secret-service/pkg/backend"
 	"github.com/keptn/keptn/secret-service/pkg/model"
 	"sync"
 )
