+++ secret-service/pkg/handler/secrethandler_test.go
@@ -4,8 +4,8 @@ import (
 	"bytes"
 	"fmt"
 	"github.com/gin-gonic/gin"
-	"github.com/keptn/keptn/secret-service/internal/backend"
-	"github.com/keptn/keptn/secret-service/internal/backend/fake"
+	"github.com/keptn/keptn/secret-service/pkg/backend"
+	"github.com/keptn/keptn/secret-service/pkg/backend/fake"
 	"github.com/keptn/keptn/secret-service/pkg/handler"
 	"github.com/keptn/keptn/secret-service/pkg/model"
 	"github.com/stretchr/testify/assert"
